Prepare codebase for desktop nesting
A websocket connection is now not in a 1:1 relation to an open desktop
anymore, rather open requests can be sent via the websocket to open
desktops. The messages emitted on the opened desktops are forwarded
to the frontend and the frontend can interact with every desktop that
it has opened.

A close request is to be implemented soon, such that desktops can
be closed again when the frontend does not need them anymore.
All desktops are closed automatically when the connection is shut
down or closed.