Allow installation of apps
This feature is nowhere near complete yet, but you can upload html
files now (file extension must be changed to ".app" first) and
install them to your local storage.

Opening the apps is not possible via the GUI, atm. but there is a
js function already which opens a window and loads the HTML into
an iframe in the window content.